+++ main.go
@@ -1,42 +1,28 @@
 package main
 
-import "github.com/gofiber/fiber/v2"
+import (
+	"net/http"
+
+	"github.com/BloomGameStudio/PuzzleService/database"
+	puzzlehttp "github.com/BloomGameStudio/PuzzleService/puzzle/http"
+	puzzlegorm "github.com/BloomGameStudio/PuzzleService/puzzle/repository"
+	puzzleuc "github.com/BloomGameStudio/PuzzleService/puzzle/usecase"
+	"github.com/gofiber/fiber/v2"
+)
 
 func main() {
-
 	// Create a new instance of the Fiber app.
 	app := fiber.New()
 
-	// Routes
-	app.Get("/ping", func(c *fiber.Ctx) error {
-		return c.SendString("Pong!")
+	// Liveness
+	app.Get("/liveness", func(c *fiber.Ctx) error {
+		return c.SendStatus(http.StatusOK)
 	})
 
-	// Routes for the puzzles
-	app.Get("/puzzles", func(c *fiber.Ctx) error {
-		return c.SendString("Unimplemented!")
-	})
-
-	app.Get("/puzzles/:id", func(c *fiber.Ctx) error {
-		return c.SendString("Unimplemented!")
-	})
-
-	app.Post("/puzzles", func(c *fiber.Ctx) error {
-		return c.SendString("Unimplemented!")
-	})
-
-	app.Patch("/puzzles", func(c *fiber.Ctx) error {
-		return c.SendString("Unimplemented!")
-	})
-
-	app.Delete("/puzzles", func(c *fiber.Ctx) error {
-		return c.SendString("Unimplemented!")
-	})
-
-	// Verification routes
-	app.Post("/verify/:puzzleID/:solution", func(c *fiber.Ctx) error {
-		return c.SendString("Unimplemented!")
-	})
+	// Puzzle
+	puzzleRepository := puzzlegorm.NewPuzzleRepository(database.GetDB())
+	puzzleUseCase := puzzleuc.NewPuzzleUseCase(puzzleRepository)
+	puzzlehttp.RegisterHTTPEndpoints(app, puzzleUseCase)
 
 	// Start the server and listen on port 3000
 	app.Listen(":3000")