+++ models/puzzle.go
@@ -1,0 +1,7 @@
+package models
+
+type Puzzle struct {
+	ID       [32]byte
+	Solution []byte
+	Title    string
+}